requirements.txt
Django==3.2.5
djangorestframework==3.12.4
PyJWT==2.1.0
django-cors-headers==3.7.0
djongo==1.3.6
python-dotenv==0.19.0